Clean unused refs
