+++ libs/agent/agent/core/loop.py
@@ -3,6 +3,7 @@
 import logging
 import asyncio
 from abc import ABC, abstractmethod
+from enum import Enum, auto
 from typing import Any, AsyncGenerator, Dict, List, Optional, Tuple
 from datetime import datetime
 
@@ -10,6 +11,14 @@ from computer import Computer
 from .experiment import ExperimentManager
 
 logger = logging.getLogger(__name__)
+
+
+class AgentLoop(Enum):
+    """Enumeration of available loop types."""
+
+    ANTHROPIC = auto()  # Anthropic implementation
+    OMNI = auto()  # OmniLoop implementation
+    # Add more loop types as needed
 
 
 class BaseLoop(ABC):
@@ -191,156 +200,3 @@ class BaseLoop(ABC):
         """
         if self.experiment_manager:
             self.experiment_manager.save_screenshot(img_base64, action_type)
-
-    def _create_openai_compatible_response(
-        self, response: Any, messages: List[Dict[str, Any]]
-    ) -> Dict[str, Any]:
-        """Create an OpenAI computer use agent compatible response format.
-
-        Args:
-            response: The original API response
-            messages: List of messages in standard OpenAI format
-
-        Returns:
-            A response formatted according to OpenAI's computer use agent standard
-        """
-        import json
-
-        # Create a unique ID for this response
-        response_id = f"resp_{datetime.now().strftime('%Y%m%d%H%M%S')}_{id(response)}"
-        reasoning_id = f"rs_{response_id}"
-        action_id = f"cu_{response_id}"
-        call_id = f"call_{response_id}"
-
-        # Extract the last assistant message
-        assistant_msg = None
-        for msg in reversed(messages):
-            if msg["role"] == "assistant":
-                assistant_msg = msg
-                break
-
-        if not assistant_msg:
-            # If no assistant message found, create a default one
-            assistant_msg = {"role": "assistant", "content": "No response available"}
-
-        # Initialize output array
-        output_items = []
-
-        # Extract reasoning and action details from the response
-        content = assistant_msg["content"]
-        reasoning_text = None
-        action_details = None
-
-        # Extract reasoning and action from different content formats
-        if isinstance(content, str):
-            try:
-                # Try to parse JSON
-                parsed_content = json.loads(content)
-                reasoning_text = parsed_content.get("Explanation", "")
-
-                # Extract action details
-                action = parsed_content.get("Action", "")
-                position = parsed_content.get("Position", {})
-                text_input = parsed_content.get("Text", "")
-
-                if action.lower() == "click" and position:
-                    action_details = {
-                        "type": "click",
-                        "button": "left",
-                        "x": position.get("x", 100),
-                        "y": position.get("y", 100),
-                    }
-                elif action.lower() == "type" and text_input:
-                    action_details = {
-                        "type": "type",
-                        "text": text_input,
-                    }
-                elif action.lower() == "scroll":
-                    action_details = {
-                        "type": "scroll",
-                        "x": 100,
-                        "y": 100,
-                        "scroll_x": position.get("delta_x", 0),
-                        "scroll_y": position.get("delta_y", 0),
-                    }
-            except json.JSONDecodeError:
-                # If not valid JSON, use the content as reasoning
-                reasoning_text = content
-        elif isinstance(content, list):
-            # Handle list of content blocks (like Anthropic format)
-            for item in content:
-                if isinstance(item, dict):
-                    if item.get("type") == "text":
-                        # Collect text blocks for reasoning
-                        if reasoning_text is None:
-                            reasoning_text = ""
-                        reasoning_text += item.get("text", "")
-                    elif item.get("type") == "tool_use":
-                        # Extract action from tool_use (similar to Anthropic format)
-                        tool_input = item.get("input", {})
-                        if "click" in tool_input or "position" in tool_input:
-                            position = tool_input.get("click", tool_input.get("position", {}))
-                            if isinstance(position, dict) and "x" in position and "y" in position:
-                                action_details = {
-                                    "type": "click",
-                                    "button": "left",
-                                    "x": position.get("x", 100),
-                                    "y": position.get("y", 100),
-                                }
-                        elif "type" in tool_input or "text" in tool_input:
-                            action_details = {
-                                "type": "type",
-                                "text": tool_input.get("type", tool_input.get("text", "")),
-                            }
-                        elif "scroll" in tool_input:
-                            scroll = tool_input.get("scroll", {})
-                            action_details = {
-                                "type": "scroll",
-                                "x": 100,
-                                "y": 100,
-                                "scroll_x": scroll.get("x", 0),
-                                "scroll_y": scroll.get("y", 0),
-                            }
-
-        # Add reasoning item if we have text content
-        if reasoning_text:
-            output_items.append(
-                {
-                    "type": "reasoning",
-                    "id": reasoning_id,
-                    "summary": [
-                        {
-                            "type": "summary_text",
-                            "text": reasoning_text[:200],  # Truncate to reasonable length
-                        }
-                    ],
-                }
-            )
-
-        # If no action details extracted, use default
-        if not action_details:
-            action_details = {
-                "type": "click",
-                "button": "left",
-                "x": 100,
-                "y": 100,
-            }
-
-        # Add computer_call item
-        computer_call = {
-            "type": "computer_call",
-            "id": action_id,
-            "call_id": call_id,
-            "action": action_details,
-            "pending_safety_checks": [],
-            "status": "completed",
-        }
-        output_items.append(computer_call)
-
-        # Create the OpenAI-compatible response format
-        return {
-            "output": output_items,
-            "id": response_id,
-            # Include the original response for compatibility
-            "response": {"choices": [{"message": assistant_msg, "finish_reason": "stop"}]},
-        }
