--- libs/agent/agent/providers/omni/tool_manager.py
@@ -1,91 +1,0 @@
-# """Omni tool manager implementation."""
-
-# from typing import Dict, List, Type, Any
-
-# from computer import Computer
-# from ...core.tools import BaseToolManager, BashTool, EditTool
-
-# class OmniToolManager(BaseToolManager):
-#     """Tool manager for multi-provider support."""
-
-#     def __init__(self, computer: Computer):
-#         """Initialize Omni tool manager.
-
-#         Args:
-#             computer: Computer instance for tools
-#         """
-#         super().__init__(computer)
-
-#     def get_anthropic_tools(self) -> List[Dict[str, Any]]:
-#         """Get tools formatted for Anthropic API.
-
-#         Returns:
-#             List of tool parameters in Anthropic format
-#         """
-#         tools: List[Dict[str, Any]] = []
-
-#         # Map base tools to Anthropic format
-#         for tool in self.tools.values():
-#             if isinstance(tool, BashTool):
-#                 tools.append({
-#                     "type": "bash_20241022",
-#                     "name": tool.name
-#                 })
-#             elif isinstance(tool, EditTool):
-#                 tools.append({
-#                     "type": "text_editor_20241022",
-#                     "name": "str_replace_editor"
-#                 })
-
-#         return tools
-
-#     def get_openai_tools(self) -> List[Dict]:
-#         """Get tools formatted for OpenAI API.
-
-#         Returns:
-#             List of tool parameters in OpenAI format
-#         """
-#         tools = []
-
-#         # Map base tools to OpenAI format
-#         for tool in self.tools.values():
-#             tools.append({
-#                 "type": "function",
-#                 "function": tool.get_schema()
-#             })
-
-#         return tools
-
-#     def get_groq_tools(self) -> List[Dict]:
-#         """Get tools formatted for Groq API.
-
-#         Returns:
-#             List of tool parameters in Groq format
-#         """
-#         tools = []
-
-#         # Map base tools to Groq format
-#         for tool in self.tools.values():
-#             tools.append({
-#                 "type": "function",
-#                 "function": tool.get_schema()
-#             })
-
-#         return tools
-
-#     def get_qwen_tools(self) -> List[Dict]:
-#         """Get tools formatted for Qwen API.
-
-#         Returns:
-#             List of tool parameters in Qwen format
-#         """
-#         tools = []
-
-#         # Map base tools to Qwen format
-#         for tool in self.tools.values():
-#             tools.append({
-#                 "type": "function",
-#                 "function": tool.get_schema()
-#             })
-
-#         return tools
