+++ libs/agent/agent/providers/omni/tools/__init__.py
@@ -1,11 +1,1 @@
 """Omni provider tools - compatible with multiple LLM providers."""
-
-from .bash import OmniBashTool
-from .computer import OmniComputerTool
-from .manager import OmniToolManager
-
-__all__ = [
-    "OmniBashTool",
-    "OmniComputerTool",
-    "OmniToolManager",
-]
